+++ src/RcppExports.cpp
@@ -249,6 +249,16 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// hybrids
+List hybrids();
+RcppExport SEXP _dplyr_hybrids() {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    rcpp_result_gen = Rcpp::wrap(hybrids());
+    return rcpp_result_gen;
+END_RCPP
+}
 // get_date_classes
 SEXP get_date_classes();
 static SEXP _dplyr_get_date_classes_try() {
@@ -736,6 +746,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_dplyr_group_size_grouped_cpp", (DL_FUNC) &_dplyr_group_size_grouped_cpp, 1},
     {"_dplyr_grouped_df_impl", (DL_FUNC) &_dplyr_grouped_df_impl, 2},
     {"_dplyr_group_data_grouped_df", (DL_FUNC) &_dplyr_group_data_grouped_df, 1},
+    {"_dplyr_hybrids", (DL_FUNC) &_dplyr_hybrids, 0},
     {"_dplyr_get_date_classes", (DL_FUNC) &_dplyr_get_date_classes, 0},
     {"_dplyr_get_time_classes", (DL_FUNC) &_dplyr_get_time_classes, 0},
     {"_dplyr_semi_join_impl", (DL_FUNC) &_dplyr_semi_join_impl, 5},
